requirements.txt
Flask==2.1.0
Flask-SQLAlchemy==3.0.0
Flask-Login==0.6.2
Werkzeug==2.3.0
langchain==0.1.0
langchain-community==0.0.10
langchain-google-genai==0.0.4
faiss-cpu==1.7.4
python-dotenv==0.19.0
PyPDF2==3.0.0
google-generativeai==0.3.0